third_party/blink/web_tests/external/wpt/app-history/currentchange-event/currentchange-app-history-reload-transitionWhile-expected.txt

This is a testharness.js-based test.
FAIL AppHistoryCurrentChangeEvent fires for appHistory.reload() intercepted by transitionWhile assert_equals: expected object "[object AppHistoryEntry]" but got object "[object AppHistoryEntry]"
Harness: the test ran to completion.
